Add Tpl to uefi-raw and use it from uefi
